+++ MarvelUI/Code/Module/Character/ComicCharacterDataIn.swift
@@ -1,0 +1,15 @@
+//
+//  ComicCharacterDataIN.swift
+//  MarvelUI
+//
+//  Created by Abby Dominguez on 5/6/24.
+//
+
+import Foundation
+
+struct CharacterDataIn: Codable {
+    let code: Int
+    let status, copyright, attributionText, attributionHTML: String
+    let etag: String
+    let data: CharacterDataClass
+}
