+++ MarvelUI/Code/Module/Comic/Comics.swift
@@ -1,0 +1,15 @@
+//
+//  Comics.swift
+//  MarvelUI
+//
+//  Created by Abby Dominguez on 5/6/24.
+//
+
+import Foundation
+
+struct Comics: Codable {
+    let available: Int
+    let collectionURI: String
+    let items: [ComicsItem]
+    let returned: Int
+}
