+++ MarvelUI/Code/Module/Comic/ComicsItem.swift
@@ -1,0 +1,13 @@
+//
+//  ComicsItem.swift
+//  MarvelUI
+//
+//  Created by Abby Dominguez on 5/6/24.
+//
+
+import Foundation
+
+struct ComicsItem: Codable {
+    let resourceURI: String
+    let name: String
+}
