+++ MarvelUI/Code/Module/Story/StoriesItem.swift
@@ -1,0 +1,14 @@
+//
+//  StoriesItem.swift
+//  MarvelUI
+//
+//  Created by Abby Dominguez on 5/6/24.
+//
+
+import Foundation
+
+struct StoriesItem: Codable {
+    let resourceURI: String
+    let name: String
+    let type: String
+}
